move user related entities to shared path
