+++ apps/server/src/modules/news/mapper/target-info.mapper.ts
@@ -1,8 +1,8 @@
-import { NewsTargetInfo } from '@shared/domain/types/news.types';
+import { NewsTarget } from '@shared/domain/types/news.types';
 import { TargetInfoResponse } from '../controller/dto/target-info.response';
 
 export class TargetInfoMapper {
-	static mapToResponse(target: NewsTargetInfo): TargetInfoResponse {
+	static mapToResponse(target: NewsTarget): TargetInfoResponse {
 		const dto = new TargetInfoResponse();
 		dto.id = target.id;
 		dto.name = target.name;
