--- apps/server/src/modules/user/entity/index.ts
@@ -1,3 +1,0 @@
-export * from './account.entity';
-export * from './user.entity';
-export * from './role.entity';
