+++ apps/server/src/modules/user/repo/role.repo.integration.spec.ts
@@ -1,8 +1,8 @@
 import { NotFoundError, ValidationError } from '@mikro-orm/core';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
+import { Role } from '@shared/domain';
 import { MongoMemoryDatabaseModule } from '../../database';
-import { Role } from '../entity';
 import { RoleRepo } from './role.repo';
 
 describe('role repo', () => {
