+++ apps/server/src/shared/domain/entity/index.ts
@@ -1,8 +1,10 @@
+export * from './account.entity';
 export * from './base.entity';
 export * from './course.entity';
 export * from './coursegroup.entity';
 export * from './dashboard.entity';
 export * from './news.entity';
+export * from './role.entity';
 export * from './school.entity';
 export * from './team.entity';
 export * from './user.entity';
