+++ apps/server/src/shared/domain/entity/news.entity.ts
@@ -1,11 +1,11 @@
 import { Entity, Enum, ManyToOne, Property } from '@mikro-orm/core';
 import { BaseEntityWithTimestamps } from './base.entity';
 import { EntityId } from '../types/entity-id';
-import { Course } from './course.entity';
-import { School } from './school.entity';
-import { Team } from './team.entity';
-import { User } from './user.entity';
-import { NewsTargetInfo, NewsTargetModel } from '../types/news.types';
+import type { Course } from './course.entity';
+import type { School } from './school.entity';
+import type { Team } from './team.entity';
+import type { User } from './user.entity';
+import { NewsTarget, NewsTargetModel } from '../types/news.types';
 
 export interface INewsProperties {
 	title: string;
@@ -48,19 +48,19 @@ export abstract class News extends BaseEntityWithTimestamps {
 	sourceDescription?: string;
 
 	/** id reference to a collection populated later with name */
-	target!: NewsTargetInfo;
+	target!: NewsTarget;
 
 	/** name of a collection which is referenced in target */
 	@Enum()
 	targetModel: NewsTargetModel;
 
-	@ManyToOne({ fieldName: 'schoolId' })
+	@ManyToOne('School', { fieldName: 'schoolId' })
 	school!: School;
 
-	@ManyToOne({ fieldName: 'creatorId' })
+	@ManyToOne('User', { fieldName: 'creatorId' })
 	creator!: User;
 
-	@ManyToOne({ fieldName: 'updaterId' })
+	@ManyToOne('User', { fieldName: 'updaterId' })
 	updater?: User;
 
 	permissions: string[] = [];
@@ -91,18 +91,18 @@ export abstract class News extends BaseEntityWithTimestamps {
 
 @Entity({ discriminatorValue: NewsTargetModel.School })
 export class SchoolNews extends News {
-	@ManyToOne()
+	@ManyToOne('School')
 	target: School;
 }
 
 @Entity({ discriminatorValue: NewsTargetModel.Course })
 export class CourseNews extends News {
-	@ManyToOne()
+	@ManyToOne('Course')
 	target: Course;
 }
 
 @Entity({ discriminatorValue: NewsTargetModel.Team })
 export class TeamNews extends News {
-	@ManyToOne()
+	@ManyToOne('Team')
 	target: Team;
 }
