+++ apps/server/src/shared/domain/entity/user.entity.ts
@@ -1,12 +1,14 @@
-import { Entity, Property, Index, Unique } from '@mikro-orm/core';
+import { Collection, Entity, ManyToMany, Property, Index, Unique } from '@mikro-orm/core';
+import type { Role } from './role.entity';
 import { EntityId } from '../types';
 import { BaseEntityWithTimestamps } from './base.entity';
 
 export interface IUserProperties {
 	email: string;
-	firstName: string;
-	lastName: string;
+	firstName?: string;
+	lastName?: string;
 	school: EntityId;
+	roles: Role[];
 }
 
 @Entity({ tableName: 'users' })
@@ -17,10 +19,14 @@ export class User extends BaseEntityWithTimestamps {
 	email: string;
 
 	@Property()
-	firstName: string;
+	firstName?: string;
 
 	@Property()
-	lastName: string;
+	lastName?: string;
+
+	@Index({ name: 'roleIdBasedSearches' })
+	@ManyToMany('Role', undefined, { fieldName: 'roles' })
+	roles = new Collection<Role>(this);
 
 	@Index({ name: 'searchUserForSchool' })
 	@Property({ fieldName: 'schoolId' })
@@ -32,7 +38,6 @@ export class User extends BaseEntityWithTimestamps {
 		this.lastName = props.lastName;
 		this.email = props.email;
 		this.school = props.school;
-
-		Object.assign(this, {});
+		this.roles = new Collection<Role>(this, props.roles || []);
 	}
 }
